Update index.js
Saves one character in the input, but does not effect the output.

Using a RegEx-escapable character (e.g. '{'), that does not need to be escaped like the backslash saves one character in the argument, but it does not effect the output, because the generated string would be '\\' prior to or '\{' after the change. Other possibilities, that work and are allowed according to the whitelisted charset would be: }[]()/+